Refactor code for Page, Today and Week Fragments (get rid of code dupplications)
